pyproject.toml
[project]
name = "agent-frontend-template"
version = "0.1.0"
description = "An example AI Agent as a template for quick setup of applications."
authors = [
    {name = "Christoph Pernul", email = "cpernul@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.85.0",
    "openai-agents[litellm]>=0.0.17",
#    Langgraph
    "langchain-anthropic>=0.3.10",
    "langchain-community>=0.3.20",
    "langchain>=1.1.0",
#    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.9",
    "langgraph>=0.3.18",
    "langgraph-checkpoint-sqlite>=2.0.6",
    "langsmith>=0.3.18",
    "pillow", # to show image of graph
#    Tools
    "playwright>=1.51.0",
    "polygon-api-client>=1.14.5",
#    "sendgrid>=6.11.0",
#    frontend
    "chainlit>=2.8.3",
    "gradio>=5.33.1",
    "httpx>=0.28.1",
    "bs4>=0.0.2",
    "requests>=2.32.4",
    "python-dotenv>=1.1.0",
    "setuptools>=80.9.0",
]



[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
    "pre-commit>=4.3.0",
]
